+++ CHANGELOG.md
@@ -1,5 +1,6 @@
-Unreleased
-----------
+0.3.3
+-----
+- Decreased `Readline` object size by heap allocating more state
 - Fixed potential memory leak caused by undo lists
 - Bumped minimum supported Rust version to `1.38`